figuring out coin count

+++ DimensionRun/Code/Map.cpp
@@ -216,19 +216,6 @@ void Map::LoadMap(const std::string& l_path) {
 			keystream >> enemyX >> enemyY;
 			entityMgr->Find(enemyId)->SetPosition(enemyX, enemyY);
 		}
-		else if (type == "COIN") {
-			int coinId = entityMgr->AddEntity(EntityType::Coin);
-
-			if (coinId < 0) {
-				continue;
-			}
-
-			float coinX = 0;
-			float coinY = 0;
-
-			keystream >> coinX >> coinY;
-			entityMgr->Find(coinId)->SetPosition(coinX, coinY);
-		}
 		else {
 			// Something else.
 			std::cout << "! Unknown type \"" << type << "\"." << std::endl;
@@ -395,7 +382,11 @@ void Map::PurgeMap() {
 		delete itr.second;
 	}
 	m_TileMap.clear();
-	m_Context->m_EntityManager->Purge();
+	//m_Context->m_EntityManager->Purge();
+	
+	if (!m_Context->m_EntityManager->Find("Player")) {
+		m_Context->m_EntityManager->Purge();
+	}
 
 	if (m_BackgroundBack == "") { 
 		return;
